Fix 55 logic and security bugs in streamIO/integer
Covers all 55 rows recorded for streamIO/integer in HANDOFF.md.

Security:
- FilterCrypt is a home-grown XOR stream cipher that presented itself as
  real encryption. Per the agreed remediation it keeps its API but is now
  @Deprecated and its Javadoc states plainly that it is not
  cryptographically secure and must not be used for confidentiality. No
  crypto library was swapped in.
- AStatement.getResultSet(String) built a File path from an unsanitised
  table name. CHECK_TABLE_NAME now rejects null/empty names, path
  separators, ':', NUL, '.', '..' and absolute paths with an SQLException
  before the File is constructed.

Correctness, main families:
- Sign and width errors: BigEndianReader shifted an int by 32 bits;
  StreamIn_Primitive.nextEnum and FilterInByBitMask shifted an int where a
  long was needed; FilterIn_Byte sign-extended a byte.
- Hex nibble conversion was wrong in both directions in FilterBinHex2Byte
  and FilterByte2BinHex ('9' instead of '0', 'A'+10 instead of 'A'-10).
- FilterCRC16/FilterCRC32 dropped a legitimate NUL byte by testing ret > 0.
- ArrayStreamIn_Int.nextLongInternal looped back into itself
  (StackOverflowError); RandomMix.reset threw instead of delegating.
- StreamOutInstantiator's six typed add* methods all sized and copied off
  the int accumulator instead of their own.
- AConnection.close leaked open Statements; it now tracks them via a
  registry that AStatement's canonical constructor populates.

Three rows were misdiagnosed by the documentation pass and were fixed to
match the actual defect, which is noted per file:
- FilterASCII2Base64: the division is correct; the real corruption came
  from stale trailing buffer bytes, now zeroed before encoding.
- Depeater.flush: the shortened length is the protocol's tail signal, not
  a bug; the genuine defect was the missing super.flush().
- RSMetaData: its column indices are 0-based, but so is the rest of the
  family (AResultSet.findColumn, FilterResultSet, ResultSetArray). Making
  this one class 1-based would have made it the odd one out, so the
  0-based contract is now documented and every accessor bounds-checked.
  Converting the whole family to the JDBC-standard 1-based indexing is a
  real but larger change and should be scheduled separately.

Also refreshed two class comments in CallStatementFix and PrepStatementFix
that still pointed at TODO markers now removed.

44 of the 48 touched files compile under javac 11. DriverFix,
CallStatementFix, PrepStatementFix and ResultSetArray fail only on the
pre-existing JDBC 4.1 gaps, confirmed by compiling their HEAD versions.
Runtime verification was not possible: the in-corpus test helper
streamIO.Assert transitively needs javax.xml.bind, absent from Java 11.

+++ streamIO/integer/adapter/ReaderToStreamIn_Byte.java
@@ -130,12 +130,7 @@ implements IStreamIn_Byte {
 			int val;
 			if ((val = reader.read()) < 0) {
 				return i-off; }
-			// TODO: LOGIC: narrowing the Character to (byte) before storing into this int[]
-			// truncates and sign-extends any code point outside -128..127 (e.g. Latin-1
-			// supplement, or any char > 0x7F), corrupting the value that ends up in b[i] -
-			// unlike the read(byte[], ...) overload above, where truncating to byte is the
-			// intended behavior, this int[] overload should store "val" directly, unnarrowed.
-			b[i] = (byte) val;
+			b[i] = val;
 		}
 		return len; }
